==> notebooks/03_classify_store_status.ipynb <==
# -----------------------------
# 1. Import libraries and files
# -----------------------------

# Import libraries
import pandas as pd

# Load business file
df_b = pd.read_csv('../datasets/business3.csv')

# Load review file
df_r = pd.read_parquet("../datasets/parquet_file/review_filtered.parquet")

# -----------------------------
# 2. Classify store_status
# -----------------------------

df_status = df_b[['business_id', 'name', 'is_open', 'stars', 'review_count']]

# Ensure date column is in datetime format
df_r['date'] = pd.to_datetime(df_r['date'])

# Calculate the earliest review date per business_id
first_review_dates = df_r.groupby('business_id')['date'].min().reset_index()
first_review_dates = first_review_dates.rename(columns={'date': 'first_review_date'})

# Merge with df_status to include earliest review date
df_status = df_status.merge(first_review_dates, on='business_id', how='left')

# Check results
df_status.head()

# Define cutoff conditions
cutoff_date = pd.to_datetime("2019-01-19")
min_review = 50  # updated threshold

# Calculate 70th percentile for average rating
rating_threshold = df_status['stars'].quantile(0.7)

# Initialize store_status column
df_status['store_status'] = 'unknown'

# Assign store_status based on conditions
# close_real
df_status.loc[
    (df_status['is_open'] == 0) &
    ((df_status['review_count'] < min_review) | (df_status['stars'] <= rating_threshold)),
    'store_status'
] = 'close_real'

# close_external
df_status.loc[
    (df_status['is_open'] == 0) &
    (df_status['store_status'] == 'unknown'),
    'store_status'
] = 'close_external'

# open_old
df_status.loc[
    (df_status['is_open'] == 1) &
    (df_status['first_review_date'] < cutoff_date),
    'store_status'
] = 'open_old'

# open_new
df_status.loc[
    (df_status['is_open'] == 1) &
    (df_status['first_review_date'] >= cutoff_date),
    'store_status'
] = 'open_new'

# Check results
df_status['store_status'].value_counts()


df_status['store_status'].value_counts()

# Merge store_status from df_status into df_b based on business_id
df_b = df_b.merge(df_status[['business_id', 'store_status']], on='business_id', how='left')

df_b.head()

df_b.to_csv('../datasets/business4.csv', index=False)
